requirements.txt
black
requests
pylint
pytest
pytest-cov
requests_testing
azure-storage-blob
azure-storage
